[Chore] #107 로그인시 필요한 정보 받아오는 로직 viewModel로 이동
